test: share POTA certificate fixture
Move reusable POTA root and PTA-chain setup into the X.509 test harness.
Keep PartFinal policy construction in the command test layer.

Copilot-Session: 7ccbdc1e-2946-4f0c-b020-b3f1ade57f77
